backend/requirements.txt
anyio==3.6.1
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
ecdsa==0.17.0
fastapi==0.78.0
filelock==3.7.0
h11==0.13.0
huggingface-hub==0.7.0
idna==3.3
motor==3.0.0
numpy==1.22.4
packaging==21.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.9
python-decouple==3.6
python-jose==3.3.0
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
tqdm==4.64.0
transformers==4.19.2
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
